Start work on separate SDK library for completely independant modules.
