+++ base/CMakeLists.txt
@@ -8,5 +8,10 @@ SET(CAER_BASE_CXX_FILES
 	base/module.cpp
 	base/mainloop.cpp)
 
+SET(LIBCAERSUPPORT_SRC_FILES
+	base/module_lib.cpp
+	base/mainloop_lib.cpp
+	PARENT_SCOPE)
+
 SET(CAER_C_SRC_FILES ${CAER_C_SRC_FILES} ${CAER_BASE_C_FILES} PARENT_SCOPE)
 SET(CAER_CXX_SRC_FILES ${CAER_CXX_SRC_FILES} ${CAER_BASE_CXX_FILES} PARENT_SCOPE)
