+++ base/config.h
@@ -1,7 +1,7 @@
 #ifndef CONFIG_H_
 #define CONFIG_H_
 
-#include "main.h"
+#include "lib/utils.h"
 
 #ifdef __cplusplus
 extern "C" {
