+++ base/config_server.cpp
@@ -1,7 +1,7 @@
 #include "config_server.h"
 #include "mainloop.h"
-#include "ext/threads_ext.h"
-#include "ext/pathmax.h"
+#include "lib/cross/threads_ext.h"
+#include "lib/cross/pathmax.h"
 
 #include <atomic>
 #include <thread>
