+++ base/mainloop.cpp
@@ -1,9 +1,8 @@
 #include "mainloop.h"
-#include "ext/pathmax.h"
+#include "lib/cross/pathmax.h"
 #include <csignal>
 
 #include <regex>
-#include <unordered_map>
 #include <unordered_set>
 #include <queue>
 #include <sstream>
@@ -36,182 +35,6 @@
 
 #include <libcaercpp/libcaer.hpp>
 using namespace libcaer::log;
-
-struct OrderedInput {
-	int16_t typeId;
-	int16_t afterModuleId;
-	bool copyNeeded;
-
-	OrderedInput(int16_t t, int16_t a) :
-			typeId(t),
-			afterModuleId(a),
-			copyNeeded(false) {
-	}
-
-	// Comparison operators.
-	bool operator==(const OrderedInput &rhs) const noexcept {
-		return (typeId == rhs.typeId);
-	}
-
-	bool operator!=(const OrderedInput &rhs) const noexcept {
-		return (typeId != rhs.typeId);
-	}
-
-	bool operator<(const OrderedInput &rhs) const noexcept {
-		return (typeId < rhs.typeId);
-	}
-
-	bool operator>(const OrderedInput &rhs) const noexcept {
-		return (typeId > rhs.typeId);
-	}
-
-	bool operator<=(const OrderedInput &rhs) const noexcept {
-		return (typeId <= rhs.typeId);
-	}
-
-	bool operator>=(const OrderedInput &rhs) const noexcept {
-		return (typeId >= rhs.typeId);
-	}
-};
-
-struct ModuleInfo {
-	// Module identification.
-	int16_t id;
-	const std::string name;
-	// SSHS configuration node.
-	sshsNode configNode;
-	// Parsed moduleInput configuration.
-	std::unordered_map<int16_t, std::vector<OrderedInput>> inputDefinition;
-	// Connectivity graph (I/O).
-	std::vector<std::pair<ssize_t, ssize_t>> inputs;
-	std::unordered_map<int16_t, ssize_t> outputs;
-	// Loadable module support.
-	const std::string library;
-	ModuleLibrary libraryHandle;
-	caerModuleInfo libraryInfo;
-	// Module runtime data.
-	caerModuleData runtimeData;
-
-	ModuleInfo() :
-			id(-1),
-			name(),
-			configNode(nullptr),
-			library(),
-			libraryHandle(),
-			libraryInfo(nullptr),
-			runtimeData(nullptr) {
-	}
-
-	ModuleInfo(int16_t i, const std::string &n, sshsNode c, const std::string &l) :
-			id(i),
-			name(n),
-			configNode(c),
-			library(l),
-			libraryHandle(),
-			libraryInfo(nullptr),
-			runtimeData(nullptr) {
-	}
-};
-
-struct DependencyNode;
-
-struct DependencyLink {
-	int16_t id;
-	std::shared_ptr<DependencyNode> next;
-
-	DependencyLink(int16_t i) :
-			id(i) {
-	}
-
-	// Comparison operators.
-	bool operator==(const DependencyLink &rhs) const noexcept {
-		return (id == rhs.id);
-	}
-
-	bool operator!=(const DependencyLink &rhs) const noexcept {
-		return (id != rhs.id);
-	}
-
-	bool operator<(const DependencyLink &rhs) const noexcept {
-		return (id < rhs.id);
-	}
-
-	bool operator>(const DependencyLink &rhs) const noexcept {
-		return (id > rhs.id);
-	}
-
-	bool operator<=(const DependencyLink &rhs) const noexcept {
-		return (id <= rhs.id);
-	}
-
-	bool operator>=(const DependencyLink &rhs) const noexcept {
-		return (id >= rhs.id);
-	}
-};
-
-struct DependencyNode {
-	size_t depth;
-	int16_t parentId;
-	DependencyNode *parentLink;
-	std::vector<DependencyLink> links;
-
-	DependencyNode(size_t d, int16_t pId, DependencyNode *pLink) :
-			depth(d),
-			parentId(pId),
-			parentLink(pLink) {
-	}
-};
-
-struct ActiveStreams {
-	int16_t sourceId;
-	int16_t typeId;
-	bool isProcessor;
-	std::vector<int16_t> users;
-	std::shared_ptr<DependencyNode> dependencies;
-
-	ActiveStreams(int16_t s, int16_t t) :
-			sourceId(s),
-			typeId(t),
-			isProcessor(false) {
-	}
-
-	// Comparison operators.
-	bool operator==(const ActiveStreams &rhs) const noexcept {
-		return (sourceId == rhs.sourceId && typeId == rhs.typeId);
-	}
-
-	bool operator!=(const ActiveStreams &rhs) const noexcept {
-		return (sourceId != rhs.sourceId || typeId != rhs.typeId);
-	}
-
-	bool operator<(const ActiveStreams &rhs) const noexcept {
-		return (sourceId < rhs.sourceId || (sourceId == rhs.sourceId && typeId < rhs.typeId));
-	}
-
-	bool operator>(const ActiveStreams &rhs) const noexcept {
-		return (sourceId > rhs.sourceId || (sourceId == rhs.sourceId && typeId > rhs.typeId));
-	}
-
-	bool operator<=(const ActiveStreams &rhs) const noexcept {
-		return (sourceId < rhs.sourceId || (sourceId == rhs.sourceId && typeId <= rhs.typeId));
-	}
-
-	bool operator>=(const ActiveStreams &rhs) const noexcept {
-		return (sourceId > rhs.sourceId || (sourceId == rhs.sourceId && typeId >= rhs.typeId));
-	}
-};
-
-static struct {
-	sshsNode configNode;
-	atomic_bool systemRunning;
-	atomic_bool running;
-	atomic_uint_fast32_t dataAvailable;
-	size_t copyCount;
-	std::unordered_map<int16_t, ModuleInfo> modules;
-	std::vector<ActiveStreams> streams;
-	std::vector<std::reference_wrapper<ModuleInfo>> globalExecution;
-	std::vector<caerEventPacketHeader> eventPackets;
-} glMainloopData;
 
 static int caerMainloopRunner();
 static void printDebugInformation();
@@ -1959,146 +1782,6 @@ static void printDebugInformation() {
 
 		for (const auto &o : m.get().outputs) {
 			log(logLevel::DEBUG, "Mainloop", " --> OUT: %d - %d", o.first, o.second);
-		}
-	}
-}
-
-void caerMainloopDataNotifyIncrease(void *p) {
-	UNUSED_ARGUMENT(p);
-
-	glMainloopData.dataAvailable.fetch_add(1, std::memory_order_release);
-}
-
-void caerMainloopDataNotifyDecrease(void *p) {
-	UNUSED_ARGUMENT(p);
-
-	// No special memory order for decrease, because the acquire load to even start running
-	// through a mainloop already synchronizes with the release store above.
-	glMainloopData.dataAvailable.fetch_sub(1, std::memory_order_relaxed);
-}
-
-bool caerMainloopModuleExists(int16_t id) {
-	return (glMainloopData.modules.count(id) == 1);
-}
-
-bool caerMainloopModuleIsType(int16_t id, enum caer_module_type type) {
-	return (glMainloopData.modules.at(id).libraryInfo->type == type);
-}
-
-bool caerMainloopStreamExists(int16_t sourceId, int16_t typeId) {
-	return (findBool(glMainloopData.streams.begin(), glMainloopData.streams.end(), ActiveStreams(sourceId, typeId)));
-}
-
-int16_t *caerMainloopGetModuleInputIDs(int16_t id, size_t *inputsSize) {
-	// If inputsSize is known, allow not passing it in.
-	if (inputsSize != nullptr) {
-		*inputsSize = 0;
-	}
-
-	// Only makes sense to be called from PROCESSORs or OUTPUTs, as INPUTs
-	// do not have inputs themselves.
-	if (caerMainloopModuleIsType(id, CAER_MODULE_INPUT)) {
-		return (nullptr);
-	}
-
-	size_t inDefSize = glMainloopData.modules.at(id).inputDefinition.size();
-
-	int16_t *inputs = (int16_t *) malloc(inDefSize * sizeof(int16_t));
-	if (inputs == nullptr) {
-		return (nullptr);
-	}
-
-	size_t idx = 0;
-	for (auto inDef : glMainloopData.modules.at(id).inputDefinition) {
-		inputs[idx++] = inDef.first;
-	}
-
-	if (inputsSize != nullptr) {
-		*inputsSize = idx;
-	}
-	return (inputs);
-}
-
-static inline caerModuleData caerMainloopGetSourceData(int16_t sourceID) {
-	caerModuleData moduleData = glMainloopData.modules.at(sourceID).runtimeData;
-	if (moduleData == nullptr) {
-		return (nullptr);
-	}
-
-	// Sources must be INPUTs or PROCESSORs.
-	if (caerMainloopModuleIsType(sourceID, CAER_MODULE_OUTPUT)) {
-		return (nullptr);
-	}
-
-	return (moduleData);
-}
-
-sshsNode caerMainloopGetSourceNode(int16_t sourceID) {
-	caerModuleData moduleData = caerMainloopGetSourceData(sourceID);
-	if (moduleData == nullptr) {
-		return (nullptr);
-	}
-
-	return (moduleData->moduleNode);
-}
-
-sshsNode caerMainloopGetSourceInfo(int16_t sourceID) {
-	caerModuleData moduleData = caerMainloopGetSourceData(sourceID);
-	if (moduleData == nullptr) {
-		return (nullptr);
-	}
-
-	// All sources should have a sub-node in SSHS called 'sourceInfo/',
-	// while they are running only (so check running and existence).
-	if (moduleData->moduleStatus == CAER_MODULE_STOPPED) {
-		return (nullptr);
-	}
-
-	if (!sshsExistsRelativeNode(moduleData->moduleNode, "sourceInfo/")) {
-		return (nullptr);
-	}
-
-	return (sshsGetRelativeNode(moduleData->moduleNode, "sourceInfo/"));
-}
-
-void *caerMainloopGetSourceState(int16_t sourceID) {
-	caerModuleData moduleData = caerMainloopGetSourceData(sourceID);
-	if (moduleData == nullptr) {
-		return (nullptr);
-	}
-
-	return (moduleData->moduleState);
-}
-
-sshsNode caerMainloopGetModuleNode(int16_t sourceID) {
-	caerModuleData moduleData = glMainloopData.modules.at(sourceID).runtimeData;
-	if (moduleData == nullptr) {
-		return (nullptr);
-	}
-
-	return (moduleData->moduleNode);
-}
-
-void caerMainloopResetInputs(int16_t sourceID) {
-	for (auto &m : glMainloopData.globalExecution) {
-		if (m.get().libraryInfo->type == CAER_MODULE_INPUT) {
-			m.get().runtimeData->doReset.store(sourceID);
-		}
-	}
-}
-
-void caerMainloopResetOutputs(int16_t sourceID) {
-	for (auto &m : glMainloopData.globalExecution) {
-		if (m.get().libraryInfo->type == CAER_MODULE_OUTPUT) {
-			m.get().runtimeData->doReset.store(sourceID);
-		}
-	}
-}
-
-void caerMainloopResetProcessors(int16_t sourceID) {
-	for (auto &m : glMainloopData.globalExecution) {
-		if (m.get().libraryInfo->type == CAER_MODULE_PROCESSOR) {
-			m.get().runtimeData->doReset.store(sourceID);
 		}
 	}
 }
