+++ base/misc.h
@@ -8,7 +8,7 @@
 #ifndef MISC_H_
 #define MISC_H_
 
-#include "main.h"
+#include "lib/utils.h"
 
 #ifdef __cplusplus
 extern "C" {
