+++ lib/cross/portable_misc.h
