+++ lib/utils.h
@@ -1,31 +1,17 @@
-/*
- * main.h
- *
- *  Created on: Oct 8, 2013
- *      Author: llongi
- */
-
-#ifndef MAIN_H_
-#define MAIN_H_
+#ifndef LIB_UTILS_H_
+#define LIB_UTILS_H_
 
 // Common includes, useful for everyone.
 #include <libcaer/libcaer.h>
 #include <libcaer/events/packetContainer.h>
-#include "ext/sshs/sshs.h"
+#include "lib/sshs/sshs.h"
 
 // Suppress unused argument warnings, if needed
 #define UNUSED_ARGUMENT(arg) (void)(arg)
 
-// Support symbol export on Windows GCC/Clang.
-#if (defined(_WIN32) || defined(__WIN32__) || defined(WIN32)) && (defined(__GNUC__) || defined(__clang__))
-#define CAER_SYMBOL_EXPORT __attribute__ ((__dllexport__))
-#else
-#define CAER_SYMBOL_EXPORT
-#endif
-
 #ifdef __cplusplus
 #include <libcaercpp/libcaer.hpp>
-#include "ext/sshs/sshs.hpp"
+#include "lib/sshs/sshs.hpp"
 #include <algorithm>
 #include <vector>
 
@@ -77,4 +63,4 @@ static bool vectorDetectDuplicates(std::vector<T> &vec) {
 
 #endif
 
-#endif /* MAIN_H_ */
+#endif /* LIB_UTILS_H_ */
